Move remote cache out of alloc.h
Consolidate the remote code into a single file.
